Reorganized TestStep by service
