+++ service/job/test/src/test/java/org/eclipse/kapua/service/job/test/RunJobUnitTest.java
@@ -14,22 +14,21 @@ package org.eclipse.kapua.service.job.test;
 
 import io.cucumber.junit.Cucumber;
 import io.cucumber.junit.CucumberOptions;
-
 import org.junit.runner.RunWith;
 
 @RunWith(Cucumber.class)
 @CucumberOptions(
         features = {
-                     "classpath:features/JobStepDefinitionService.feature",
-                     "classpath:features/JobService.feature"
-                   },
+                "classpath:features/JobStepDefinitionService.feature",
+                "classpath:features/JobService.feature"
+        },
         glue = {"org.eclipse.kapua.service.job.test",
                 "org.eclipse.kapua.service.job.steps",
                 "org.eclipse.kapua.qa.common"
-               },
-        plugin = { "pretty",
-                   "html:target/cucumber/JobService",
-                   "json:target/JobService_cucumber.json" },
+        },
+        plugin = {"pretty",
+                "html:target/cucumber/JobService",
+                "json:target/JobService_cucumber.json"},
         monochrome = true)
 public class RunJobUnitTest {
 }
